Clarify Hosts discrepancy with WLB bug (FQDN)
WinLogBeat has been fixed to append the .org.local to the host.name field. In the past it didn't do that so this script can be customized to remove the .org.local part of the host name or leave it, depending on the use case for the SIEM App (Hosts)

If you want to see last seen for computer1.org.local for WinLogBeat and Nessus you just need to make sure they have the same host.name field. Nessus uses the FQDN where as WinLogBeat older than 7.6.0 does not. So to much you can toggle that setting to remove the FQDN to be consistent.