Automate script for easy ingestion
This script is designed to take a folder of .nessus files and mark them as .processed after they are sent for ingest so when the script is run again it won't process already processed files. It has some basic checks to ensure that a file isn't processed twice (checks hash of file).