Archive only when destination folder is configured
This is an enhancement for user who run the same scans every day, week or month and they don't want to move the scan to a different folder.

ArchiveFolderName is no long mandatory so when not configured scans will not move in the Nessus platform.

Scans that are retrieved from the Nessus API are always the latest scan in the history. For example, if the same scan was run 5 times, the script will only download the most recent scan.